adding webfonts via npm, some fixes, and removing old stuff
