+++ .eslintrc
@@ -9,5 +9,10 @@
     "commonjs": true,
     "es6": true,
     "jquery": true
+  },
+
+  "rules": {
+    "import/prefer-default-export": 0,
+    "allowImportExportEverywhere": true
   }
 }
