+++ package.json
@@ -9,9 +9,8 @@
     "__eslintTheme": "eslint wp-content/themes/inf_theme/skin/assets/",
     "__stylelintTheme": "stylelint \"wp-content/themes/inf_theme/**/*.scss\" --syntax scss",
     "precommitCss": "npm run __stylelintTheme",
-    "precommitCssUnused": "npm run __sassUnusedTheme",
     "precommitJs": "npm run __eslintTheme",
-    "precommit": "npm run precommitJs && npm run precommitCss && npm run precommitCssUnused",
+    "precommit": "npm run precommitJs && npm run precommitCss",
     "rename": "./bin/rename.js",
     "start": "webpack --progress --watch --display-error-details --display-reasons",
     "build": "NODE_ENV=production webpack --progress"
@@ -32,7 +31,6 @@
     "eslint": "^4.19.1",
     "eslint-plugin-import": "^2.10.0",
     "expose-loader": "^0.7.5",
-    "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.11",
     "imports-loader": "^0.8.0",
     "mini-css-extract-plugin": "^0.4.0",
@@ -54,14 +52,13 @@
     "css-mqpacker": "^6.0.2",
     "cssnano": "^3.10.0",
     "jquery": "^3.3.1",
-    "jquery-match-height": "^0.7.2",
     "layzr.js": "^2.2.2",
     "media-blender": "^2.1.0",
     "normalize.css": "^8.0.0",
     "postcss-font-magician": "^2.1.1",
     "prompt-sync": "^4.1.6",
     "replace-in-file": "^3.4.0",
-    "slick-carousel": "^1.8.1",
+    "webfontloader": "^1.6.28",
     "whatwg-fetch": "^2.0.4"
   },
   "browserslist": [
