+++ wp-content/themes/inf_theme/skin/assets/scripts/fonts.js
@@ -1,5 +1,7 @@
 /* global WebFont */
 
+import WebFont from 'webfontloader';
+
 WebFont.load({
   google: {
     families: ['Open Sans:400:latin-ext'],
