+++ wp-content/themes/inf_theme/skin/assets/scripts/index.js
@@ -3,10 +3,7 @@ import 'babel-polyfill';
 import 'whatwg-fetch';
 import './bugsnag-js';
 
-import './lazyloading';
 import './fonts';
 
 import './form-fields';
 import './components';
-
-import './cookies-notification';
