+++ wp-content/themes/inf_theme/skin/assets/styles/application.scss
@@ -2,7 +2,6 @@
 // Vendors - node modules
 @import '~normalize.css';
 @import '~media-blender';
-@import '~slick-carousel/slick/slick.scss';
 
 // Utils
 @import 'utils/colors';
